[FIX] bom, material, part 이벤트 구독 수정

+++ src/main/java/com/sampoom/factory/api/bom/entity/BomMaterialProjection.java
@@ -34,9 +34,9 @@ public class BomMaterialProjection {
     private String unit;
 
     @Column(nullable = false)
-    private Integer quantity;
+    private Double quantity;                     // Integer에서 Double로 변경
 
-    public BomMaterialProjection updateFromEvent(String materialName, String materialCode, String unit, Integer quantity) {
+    public BomMaterialProjection updateFromEvent(String materialName, String materialCode, String unit, Double quantity) {
         return this.toBuilder()
                 .materialName(materialName)
                 .materialCode(materialCode)
